refactor: add info message for `--branch`
This option only works when users have
local branches that are set up to track
remotes.

The user may be confused if they are trying
to view a branch that only exists on the remote
and not locally.

This commit adds a useful info message to
inform the user that this branch is not
tracked locally and instead is defaulting
to the default remote.